refractor classroom subsection select
